# Fix performance regression of mesh map generation