--- a/src/coreComponents/mesh/CellElementRegion.hpp
+++ b/src/coreComponents/mesh/CellElementRegion.hpp
@@ -66,14 +66,14 @@ public:
    * @brief The key name for the FaceElementRegion in the object catalog.
    * @return A string containing the key name.
    */
-  static const string catalogName()
+  static string catalogName()
   { return "CellElementRegion"; }
 
   /**
    * @copydoc catalogName()
    */
-  virtual const string getCatalogName() const override final
-  { return CellElementRegion::catalogName(); }
+  virtual string getCatalogName() const override final
+  { return catalogName(); }
 
   ///@}