--- a/src/coreComponents/mesh/FaceElementSubRegion.hpp
+++ b/src/coreComponents/mesh/FaceElementSubRegion.hpp
@@ -48,16 +48,16 @@ public:
    * @brief Get catalog name.
    * @return the catalog name
    */
-  static const string catalogName()
+  static string catalogName()
   { return "FaceElementSubRegion"; }
 
   /**
    * @brief Get catalog name.
    * @return the catalog name
    */
-  virtual const string getCatalogName() const override
+  virtual string getCatalogName() const override
   {
-    return FaceElementSubRegion::catalogName();
+    return catalogName();
   }
 
   ///@}