This PR fixes performance regression of mesh map generation by:
  * removing dynamic allocations in cell-face loop
  * restoring some parallel loops that were made serial
  * removing huge over-allocations when constructing `byLowestNode` arrays (this affects memory footprint more than speed)

In addition, it restores strict total ordering for `NodesAndElementOfFace` entries, which is required to have uniquely defined (race-condition-free) maps in case of multi-threaded construction.

Timings (inclusive) for `geosx::CellBlockManager::buildMaps`, measured on an `InternalMesh` with 400k elements:

| # threads | This branch | `develop` (`9f15819`) |
| ------------ | ------------: | -----------------: |
| 1 | 0.539594 s | 28.683256 s |
| 2 | 0.328936 s | 28.594731 s |
| 4 | 0.213764 s | 28.285016 s |
| 8 | 0.171741 s | 27.934983 s |

Mesh with 5M elements:
| # threads | This branch |
| ------------ | ------------: |
| 1 | 6.532824 s |
| 2 | 3.741959 s |
| 4 | 2.423015 s |
| 8 | 1.966910 s |

(only current PR is measured, I was unable to run meshes larger than ~500k with current `develop` as it hangs indefinitely).

High watermark of mesh generation is about 2GB/1M cells (including intentional overallocations in the maps for fracture generation). I don't think we can get significantly lower than that without specialization for cell shapes.

LvArray PR: https://github.com/GEOSX/LvArray/pull/256
Rebaseline PR: https://github.com/GEOSX/integratedTests/pull/211

Rebaseline required for some of the tests (those using multiple cell blocks like "staircase" mesh, or any external meshes). No numeric diffs, but there are some changes in the order of mesh maps.