@TotoGaz 2022-03-25
> And when did the regressions come in?

It's most likely my PR #1418 which in the very beginning was aiming at dealing with arbitrary cells
So I broke some optimization to make it more generic.
This was surely a wrong idea, my bad. 😞