@klevzoff 2022-03-26
> I haven't had a chance to look at this yet but the numbers speak for themselves. Is most of this other than the rewriting of the construction of `byLowestNode` just fixing regressions?

95% of performance or more was lost in dynamic allocations occurring in `faceByLowestNode` construction - I restored the non-allocating API for `CellBlock::getFaceNodes()` and got rid of storing all face nodes in the struct. In a couple other places I replaced serial algorithms with parallel (although most of them weren't bottlenecks). In all cases when building `ArrayOfArrays`, I replaced `resizeValues()`+`appendArray()` sequence of calls with a single `resizeFromCapacities()`, and got rid of arbitrary overallocations by always performing exact capacity precomputation loop.

Everything else more or less is just cleanup/refactoring, extracting repeating patterns (such as `forEqualRanges` that comes up every time we iterate over `byLowestNode` data), etc. The diff is a bit noisier than it could be because of it.