@klevzoff 2022-03-26
> It's most likely my PR #1418 which in the very beginning was aiming at dealing with arbitrary cells
> So I broke some optimization to make it more generic.
> This was surely a wrong idea, my bad.

To be fair, a few of us (myself included) have looked at that PR and didn't spot any problems. It's only due to @jeannepellerin's work that I decided to take another look at status quo. I hope that this will provide a more meaningful baseline when measuring the effects of cell shape specialization against the general case.